+++ AlphaFS/Filesystem/Enumerations/GetFullPathOptions.cs
@@ -42,7 +42,7 @@ namespace Alphaleonis.Win32.Filesystem
       /// <summary>Return full path as long full path (Unicode format). Not valid for <see cref="Path.GetRegularPath"/>.</summary>
       AsLongPath = 8,
 
-      /// <summary>Prevents and exception from being thrown if a filesystem object does not exist. Not valid for <see cref="Path.GetRegularPath"/>.</summary>
+      /// <summary>Prevents any exception from being thrown if a filesystem object does not exist. Not valid for <see cref="Path.GetRegularPath"/>.</summary>
       ContinueOnNonExist = 16,
 
       /// <summary>Check that the path contains only valid path-characters.</summary>